major refactor + added bash script
