+++ src/faz/client/cpp_client/faz_client.cpp
@@ -4,7 +4,7 @@
 #include <string>
 #include <vector>
 
-#include "../../caw/functions/caw_functions.h"
+#include "../../../caw/functions/caw_functions.h"
 
 Status FazClient::Event(Any request, Any *response, int eventType) {
   // Setting up request/response wrappers
